Add support for saving/loading messages in lists.
Messages (TAG_MESSAGE, message ID) can already be saved as list nodes
however recompiling the kod files can cause message IDs to change. This
change will cause the string value of the message name to be saved
directly after a list node containing TAG_MESSAGE, and LoadGameListNodes
will attempt to retrieve the name when it encounters a TAG_MESSAGE list
node. The ID of the loaded message name is placed in the data for the
list node, overwriting the existing message ID.

Obviously saving a game with a message-containing list node before
applying this change and loading it afterwards will not correctly load
the saved game.