Fix rare message parameter related compiler bug.
Duplicate parameters were being checked using the current message ID for
comparison, however as the parameters are parsed before the message
header, these message IDs referred to the previous message.

This works fine nearly all the time (and the message ID isn't used for
anything else by the parameters) however if a set of parameters is being
parsed and the previous message in that class has a copy in another
(parsed in this batch) class followed by a message containing a
parameter with the same name, the compiler will return a false positive
for a duplicate parameter.

This change removes the duplicate check during parameter parsing and
moves it to the parameter list sorting.