Fix errors in pre/post increment/decrement handling.
The existing code worked for simple calculations however trying to use
the prefix/postfix operators in complex statements where the result
(from a calc on a property) needs to be stored in a temp local var, the
mismatch here between source and dest types was causing errors.

Briefly, prefix ops should perform the operation then if the source is
not the destination, store the result in both places (opcode.dest gives
dest type, source1 gives source type, opnode->source gives source ID,
opnode->dest gives dest ID). Postfix ops should store the initial value
in the destination if not equal to the source location, perform the op
then store the result in the source location (basically a cheaper +/-1).

All existing operator usage works (have to use it on a property as part
of a larger calculation to trigger the error).