Fix overlay drawing for new helm.
This handles a missing case in the overlay system - overlays that should
be drawn on top of a base item, but underneath underlays drawn on other
overlays on the base item.

Software renderer handles this by drawing the new type of overlay before
any other overlays are drawn on the base item, hardware does the same
but also has to set a Z bias equal to that of the base item to avoid
rendering the overlay in the wrong place.

Reproduced the kod hotspot index constants in object3d.h, HS_HELM is
used in FindHotspotPdib rather than using '2'.