Change linked-list of timers into min binary heap.
Timers are currently kept in a linked list, unfortunately the timers
generally have a cluster at the beginning of the list (mob movement)
so every insert/delete has to traverse minimum 30-100 elements before
potentially finding the correct place to modify.

Changing to min-heap means that worst-case scenario, 10 timers have
to be looked at (generally it is much better than this).

On my system, CreateTimer drops from 1us to 0.350us, DeleteTimer
from 1.1us to 0.6us and GetTimeRemaining from 8.6us to 1.5us.