Remove rods.
Hold/blind/purge/dazzle "break" rods removed completely - the
implementation was poor (cycle through every item the player has
checking if it's a rod) and there was no chance of them being remotely
balanced.

Damage, shield and heal rods no longer drop in-game, and the
check-every-item code for damage/shield rods has been removed. It's
possible for all three of these to be balanced (and implemented
correctly) later on, so the code remains.