Add heal rods back to tres table.
Rods are now dropped by mummies (2% chance), black spiders (1% chance)
and trolls (1% chance).

Safe rooms no longer recharge rods by default - can be turned on/off
with setting.

Max number of rods that can be held is also now a setting (default 5).

Shortened a classvar that was nearly as long as the resource it referred
to.