Add new CV room to west of main castle room.
Attached Mark's CV room to the west doors in Castle Victoria. Added a
couple braziers and lowered the light level, and rotated the room so
that it runs east-west instead of north-south. Adjusted the position
of the columns to make the room symmetrical.

Adjusted some room constants that were incorrectly placed.