Convert AMA, Sandstorm, Truce and Winds back to room enchantments.
These spells were converted to radius enchantments, but still operate
as room enchantments. The correct solution here was to improve the
RoomEnchantment class so that the spells correctly check safety etc.,
which i've done here (and in previous work with Heat and Shal/Qorbane).

These spells and others could potentially work as radius enchantments
but that class still needs a lot of work and has bugs that need to be
ironed out first. Better client-based visual indicators would also
be necessary if a radius effect is used.

Added a property to AMA to adjust its effectiveness live.