Use strings instead of rsc IDs for attack messages.
With damage numbers on room enchantments such as heat and sandstorm,
it's now possible for the game to attempt to send damage text to a user
containing the resource ID of a player that has gone offline (thus using
an ID they don't have). Sending the strings here instead will always
correctly display user names.