Add ability school/level to char create list boxes.
Char creation will now list spells by school and level (and
alphabetically within those categories) to make choosing spells a little
easier. Skills are also listed by school for when the potential rogue
school is added, though school is not sent by the server for skills
currently.

Added a new list to System to keep track of the new-player eligible
spells, to streamline sending this list.