Add email field to blakserv accounts.
Accounts must now be created with an email field. Email validity is not
checked, except to confirm that the : character is not present, and the
@ character is.

Emails can be added to existing accounts using the "set account email
acct_num email_str" admin command, and all accounts associated with an
email can be listed with the "show email email_str" admin command.

Fixed a rare issue with account deletion.

Saves with emails are backwards compatible (but email field will
obviously be deleted).