Add ability to set token generation time and scandal frequency.
Added a new property in TokenGame, piScandalFactor, which determines how
frequent scandals are. By default this is set to 100 which is the
current chance for a scandal to happen (i.e. roll is done against
Random(1,piScandalFactor). Setting this higher will decrease the chance
of a scandal.

Also added a message to set the token generation time,
SetTokenSpawnHours. This is called with the parameter iHours, which is
bound between 2 and 16, and sets each token to spawn at around that rate
(randomized slightly so they aren't all coming out at the same time.