Re-add Justicar voting system, turned off by default.
This system should never have been removed, but was done at a time when
there were no experienced contributors checking gameplay changes. The
system does have its problems (currently impossible to guarantee one
vote per person) but this might not always be the case.

The voting system is off by default but can be switched on in which case
a vote will be triggered, sign updated etc. Switching it off will remove
the Justicar immediately.

Name change and bribe systems are not affected by this, apart from the
"Justicar vote" sign not listing the bribe details (to be added).