Add option for number of users on auto acct creation.
The admin command "create auto name pw email" is a faster way to create
an account with a configurable number of user slots. Currently this just
defaults to 1, but now a config option has been added (default 3) so
this can be used for web account creation.

This number can be changed using the config option NumSlots under
[Account] in blakserv.cfg. The system won't try to create an account
with more than 10 user slots on it.

Note that the command for this is "create auto" and not "create account
auto". AdminCreateAccount will check for this (for auto and automated
anyway) and redirect the command.

Also set the default for GoPlayer and GetPlayer rights to DM level
rather than Admin level, so DMs can use these by default.