Fix Karahol's/Purge interactions, Bonk error message.
Kara'hol's Curse was holding the caster every time they were purged,
even if the spell was not removed as the
RemoveEnchantment/StartEnchantment calls are used to re-cast the spell.
This is now fixed - if the spell doesn't expire in a way that is shown
to the user, they won't be held. Purge will also no longer remove/readd
spells if no spellpower was taken.

Fixed an error message generated by Bonk on some targets.