Update ancient trinket descriptions, give time remaining in days.
Ancient trinkets will now show the time remaining in real-life days
(default 40, or 2 Meridian years). Countdown is by day and handled by
the RealTime class, so all trinkets won't expire at the same time.

Updated trinket description to let players know where to take them and
what the purpose is.

Added a FixToken() message to the StatsResetToken class (trinket) to
correctly set all trinkets to the proper time, as the property will
otherwise be set to $ when this change goes live. Update trinkets with
the command "send c statsresettoken fixtoken".