Add setting for troops attacking non-shielded factioned players.
Added a setting controlling whether faction NPC troops will attack
non-shielded players of opposing factions. A lot of players deaths by
monster are due to the higher level soldiers picking off low HP players.
While this adds some danger to the game, faction troops are prevalent
enough to make it more frustrating than enjoyable (according to
players). Considered making this level-dependent, but the differing
behavior might lead to further annoyance so the setting will switch it
completely on or off. Shielded players are always attacked.

Removed the setting for toggling Jasper vaults as it is unlikely to ever
be used. Removed the group-tougher setting as it is no longer relevant
with the XP system removing the percentage chance tougher roll.