Modify Clientpatch to put exe files at end of cache file.
Clientpatch will now put the executable files at the end of the
generated cache file so that the updater programs patch those files
last. This means if there are many files to download and the user aborts
before the end, there is less chance of breaking the exe or ending up
with an up-to-date exe but with other outdated files.

Also incremented the client version to 5042 as the installers are at
5041 (to take advantage of the smoother client update protocol given to
clients with version >= 5041).