Improve purge timer checking.
Personal enchantment timers (the first list element in the enchantment)
can hold either a valid timer or $ (infinite duration). Currently
infinite duration enchantments are skipped when purge may want to modify
them. Move the IsTimer check and verify that we don't have an infinite
duration spell first, and if anything else is present log a debug
message and skip the spell.