Fix portal entry bugs, clean up portal files.
Portals have always had an issue where players would get "stuck" on the
edge of the portal while trying to enter it. This occurred because only
the big row/col were checked for teleporting the player, meaning that
depending where the portal was placed, a player could get close enough
to not be able to move again but also not be teleported.

Portals will now call IsUserInArea in room.kod to determine if the
player is within 1 row/col of the portal in any direction and if so,
the player will be teleported. This area is pre-calculated when the
portal is placed or moved.

Cleaned up some code in portal-related files.