Adjust return type of FindListElem, remove unused check.
The NIL check at the end of FindListElem would never be true as
ret_val.int_val is already set to non-zero above, so it can be removed.

Adjusted the return type for a $ (empty) list from $ to KOD_FALSE (tag =
TAG_INT, data = 0). The rationale here is that a $ return denotes an
actual error (invalid data) but an empty list shouldn't fulfil that
condition. If the list being checked could possibly be empty and the
required action is different than the element not being found in the
list, this should be handled separately in kod instead of checking the
return value of FindListElem for two different types ($ and numeric).

Changed some FindListElem calls which were reliant on the $ return. This
also fixes a bug in BossRoom where mobs wouldn't despawn because
FindListElem was being checked for $ instead of 0 - if the list exists
and the mob isn't in it, FindListElem returns 0 (and this is why
multiple return types isn't a good idea).