Load srv strings using .rsb file, compare cli/svr rsb hash.
Server will now use the .rsb file by default for loading resources.
Building the kod files will now create the rsb in the server folder and
copy it to the client. The .rsc files are not removed from the server
directory after build as they are required for recreating the .rsb file
in the event of an incremental kod build.

The server will also store a 16 byte MD5 hash of the rsb file. The
client will create a hash of its rsb file before connecting to the
server which is sent and compared with the server's known good hash.
If there is a mismatch, a client update will be triggered (for classic
and Ogre).