Fix newbie portal move coords, fix intra-room teleporting.
Newbie portal was moving players in the wrong direction after warning
and triggering a loop of trying to warn the same player.

Added a short timer on portals to prevent them trying to teleport the
same player multiple times due to how SomethingMoved and portals work.
Also post the TeleportSomething message so the user's new coordinates
are set after the old ones (allows intra-room portals).