Set position and focus correctly for minimised dialog boxes.
If a dialog box is created while the client is minimised, the box will
stay minimised (and be unavailable to the user) when the client is
resized.

Intercept the WM_ACTIVATE message to place the dialog box correctly, and
WM_SETFOCUS + WM_WINDOWPOSCHANGING messages to return focus at the
correct time.