Set hit side effect spells as item-cast to prevent imps/reg return.
If players are morphed as mobs with hit side effects, they can obtain
reagent return for the spells by hitting monsters with the effect. The
itemcast boolean will now be set for these as it blocks reg return and
also potential improvements (if they have the spell for the side
effect).