Rotate textures by UV coords rather than on load, add mipmaps.
Floor/ceiling/wall textures need to be rotated and currently this is
done when the dx9 texture is created. This can instead be done when the
UV coordinates are calculated, and has the benefit of consolidating the
two BGF texture creation functions. This will also make implementing png
room textures easier as they can be used in the rotated state without
extra processing.

Added 5 levels of mipmaps for loaded bgfs.