Add more render options to the particle system, snow texture.
Set up the particle rendering system to use either a texture or the
existing point/lines. The particle rendering needed to be moved to after
object rendering so that the objects will be shown through any particle
transparency.

Added a snow texture (from Ogre client) and lowered the amount of snow
particles generated by 2/3 (the texture is bigger, less needed). The
additional info for the snow chunks are generated when the particle
system is initialised (or reset) rather than doing the same calcs
repeatedly for each particle. Overall the textured snow is faster than
the point snow (more work per particle, but much less of them).

Calculate the player rotation matrix once at the start of each frame and
store it rather than recalculating multiple times.