Add new island food item, dropped by avars.
Added a new island food type - "telo fruit". The item is referenced
already in a couple of ornamental items so was a good candidate for
Ko'catan-native food source. Avars, avar shamans and avar chieftains had
the item added to their loot tables. Telo fruit give 40 vigor and fill
the same as an inky-cap, and sell for 150sh average.