Consolidate guest/newb classes into one, renaming.
Consolidated the guest*.kod and newb*.kod rooms into a single set, named
for the appropriate Raza screen. Classes, resources and RIDs also
renamed, as are the Raza NPCs.

Also removed a lot of the unused/no longer needed effects dependent on
the newbie/guest room set. The three effects we might still want to
handle are turning the tutorial off, karma on mob kills in the area and
where newbies go when they die.