Add Raza as an elusion location.
Added Raza as an elusion location.

Players require the same spellpower to hit all locations as before (84
sp), in effect giving them an extra location choice for a given
spellpower.

84sp is still required to move between regions (to/from Island).

It is now slightly easier to elusion to a guildhall, as this requires
the same 5 locations as before (so 60sp now vs 70sp before). Kept it at
5 locations to very slightly improve the worth of having a guildhall.