Move goto offset adjustment from interpreter to compiler.
Goto instructions in the kod interpreter require an extra subtraction
and one or two adds to account for data being read in past the start of
the instruction (i.e. the offsets are computed based on that start
point).

Rather than handle this in the interpreter, the compiler can compute
this adjustment when outputting the instructions so that the goto
offsets are based on the origin point of when they are used by the
interpreter.