Refactor kod decompiler.
Refactored the kod decompiler (blakdeco) to work with the new BOF
format.

Added a check for correct BOF version, and exit program if this (or the
magic number) is incorrect.
Fixed a char array overrun error in str_constant().
Fixed the handling of pre/post increment/decrement opcodes (these can
perform the equivalent of two instructions and should list them both).

Removed the DEBUG_LINE opcode type - this doesn't appear to be used
anywhere in compiler/server.