Refactor kod interpreter to use dispatch over switch.
Kod interpreter now uses function dispatch for handling opcodes.
With the misc other changes to the opcodes (see compiler commit details)
this results in 10-30% better performance, and allows further
improvements to be made more easily (a giant switch statement was also
tested, and was slower than this implementation).

Added specialised versions of RetrieveValue and StoreValue for use when
the required type is known (e.g. StoreLocal). Would be cleaner if these
were inlined into the parent function, but this results in slower code
(likely the compiler won't inline into an already inlined function
correctly).

Added an admin window command to show the count for each type of opcode
executed, but disabled it in the live code as it provides unnecessary
overhead. Useful for seeing which opcodes could be optimised further
versus those which are very rarely run.

Added some null checks in database.c, unlikely to be hit but makes the
static analysis report readable.