Fix UW/feign death issues.
- Occasionally a DeadBody object will be left in a list after deletion
and get caught in garbage collection when the list is deleted. Remove
the DeadBody from FD's state when deleted.
- Fix mismatches in UW code when checking the corpse list for players.
It is possible for players to have a different name upon death (e.g.
anonymity) which then left the code unable to match to the player's real
name. DeadBody has the player's object ID, so use this instead.