Refactor phase/spectate/logging-minion interactions.
When a player is phased or spectated in or out, and remains in the room
(i.e. does not lose the effect via logging off), a message similar to
SomethingLeft will be sent out to all active objects in the room
informing them of the action.

Users handle these new messages by displaying the appropriate
phase/spectate string, and handling minions as necessary. Minion
handling logic removed from the spells themselves.
Monsters use the messages for two things: apparitions delete themselves
if caster or original target phases/spectates, and mobs that are still
under the seduce enchantment are recaptured if the owner comes back.

Seduced mobs are now set free when the owner logs off/phases/spectates,
but as noted above can be recaptured if the enchantment is still active.

Animated mobs are now dispelled if the owner leaves via
logging/phase/spectate.