Add store check on kod function calls.
New field in function table determines the function call must store the
result. Added to avoid bugs/unnecessary calls where a call (e.g.
Length(list)) is used but the result not assigned to anything. Fixed
several instances in code where call results were not stored.

Added codegen_warning to display the message to the user, but not halt
codegen.