Disable PvP and hall entry for no-pvp players.
Added SetPlayerPvPPermaOff and SetPlayerPvPPermaOn messages to turn pvp
off or on for a player. Setting it off makes the player drop faction
shield if they have one.

Non-pvp players are unable to attack or be attacked (assassin's blades
still work, as does attacking EC).
Non-pvp players are unable to enter guild halls.