Move call parm checks to compiler.
Since the call opcodes only run bytecode generated by the compiler, the
compiler knows ahead of time whether there are too many
parameters/arguments to a call. The compiler will now generate an error
if this happens, and a warning when the total number of locals is
approaching the limit (note, local limit could be raised at some point).

The parm/argument checks on the call opcodes in the interpreter have
been removed as code that would trigger these can no longer be generated
by the compiler. Gives ~5% speed increase testing on the First() C call.