Fix Silence casting, disable spell.
Silence no longer turns players outlaw on cast (which left the spell not
casting at all after recent attack code changes). Instead, spells and
communication are only blocked for valid targets of the caster, with the
outlaw effect being applied then (if necessary).

Unfortunately fixing the bugs with the spell leave us with two options:
flag outlaw on cast and affect everyone, or affect only attackables and
have the spell be wildly overpowered. As such, the spell is disabled
pending a rewrite.