Allow minions to teleport in/out of arena correctly.
UtilGoToSquare now allows mobs to bypass the ReqSomethingMoved check to
allow intra-screen teleporting where they would otherwise not be allowed
to move. TosArena uses the new parameter to allow teleporting mobs
between the arena and stands correctly, and handles teleporting minions
(instead of TosWatcher).

TosWatcher's IsCombatant now calls IsCombatant on the master for
minions, instead of checking whether the master is in the play area
(coordinate updates are posted so InPlay can be incorrect, whereas
plCombatants should be correct).