Add DM detect invisible spell.
Added a new spell for DMs, greater detect invisible. Spell is in Kraanan
(level 6) to avoid issues with a DM school PE (no timer etc).

Casting the spell on a player grants them DI for 24 hours (default time,
can be changed in the spell properties). Casting again on a player
removes the enchantment.

DMs can also use the commands "dm detect invis list" and "dm detect
invis remove playername" to show who has the enchantment and to remove
it without having access to the player (logged off or to save a
teleport).