Add deliverance touch spell for DMs.
Useful for the 'dodgebard' event, as an alternative to the targeted
spell. Works like any other touch spell, except it casts deliverance
instead of doing damage. Has a 25% chance to hit regardless of
defense/offense (can be modified in the spell's properties).

Added a success message displayed to the DM when deliverance succeeds.