Add DelLastListElem C call.
DelLastListElem removes the last element of a list, returning either the
first list node or $ if the first was the only node.

Added because DelListElem(list,Last(list)) is used in several places but
won't remove the last element if there are multiple instances if that
element (i.e. list of non-unique numbers). Also removes the need to
traverse a list twice to remove the last element in unique lists that
don't have the previous problem.

Converted DelListElem to require storing the return value since removing
the first element of a list changes the value.