DelLastListElem changes, fix unstored DelListElem call.
In news.kod, DelListElem modifies a list but the return value is not
stored for the list. This should never be an issue with this particular
use of the call (shouldn't be called with Length(list) == 1) it is still
'unsafe' so the call now reassigns the list ID to the property.

Several DelListElem(list,Last(list)) instances changed for the new
DelLastListElem(list) call for either correctness (necbal.kod) or speed
(library.kod).