Add DM spell to enchant items with itematts, add necro faction itematt.
New DM spell, can be cast on any valid item (i.e. no regs/food) and
gives a focus time for the DM to say the name of an item attribute
(paralyzer etc.) which is then applied to the item. Valid itematts
for the targeted item are listed when the spell is cast.

Added the necromancer faction itematt matching the existing duke etc
bonus damage itematts.