Fix icky cave exit, slope in outside bow hut map.
Icky cave is difficult to exit in ogre client due to the small space
between the southern bounding box of the screen and the wall. Made this
bigger and set the walls to non-passthrough to stop ogre bouncing out of
the invalid sector behind the wall.

In the outside bow hut map, it's possible to get stuck behind a ledge on
the left-hand side of the screen. Raised the height of the ledge so
players can walk out of it.