Reorder events during phase/spectate removal.
Post the SomethingXIn message to room so those effects are handled after
the enchantment is completely removed from the player phasing or
spectating back in. Send rather than post room enchantment effects on
players phasing back in as this is now called after the phase/spectate
enchantment is removed.

Added a check to make sure player is still in the room.