Check for blank itematt info desc before adding header.
If an item has item attributes but no other info description, it should
check that the itematt has an info desc before adding the special
properties header. This is to prevent items with only blank itematt
descriptions from ending up with an empty special properties section,
containing just the header.