Disallow living statues from wearing helms.
Living statues do not have equipment code implemented correctly or
completely, and items they use are not set to have any owner (and do not
appear in plPassive). This is an issue now as helms have been fixed to
ask their owner for a gender to display the correct bgf, but cannot do
so when used by statues.

Disable helms being worn by living statues until the living statue code
can be finished.