Fix misc issues.
Spellitems not used when generating loot in TreasureType need to be
deleted as they are created with object ref/lists that need to be
cleaned up properly.

Spell needs a IsAffectedByRadiusEnchantment stub to handle code that
will call IsAffectedByRadiusEnchantment on anything that can cause
damage.

Fixed a typo in stat reset, checking against wrong myst value.