Add new goto opcodes for $ checks.
Many goto true/false checks are comparisons against $ (equals or not
equals) and currently these are checked by comparing the values,
assigning the result (kod true or false) to a temp variable and then
performing a GOTO_IF_TRUE or GOTO_IF_FALSE on that temp var.

This can be simplified by performing the $ check and the goto in one
opcode, making an if (var = $) check about 20% faster. Foreach loops
also use a null check for the list, and are 20% faster if the list is $,
with the overall foreach boilerplate (i.e. not the kod code inside the
loop) about 40% faster for a non-$ list.

This has also been enabled for while loops, do-while loops and for loop
conditions. Only single statements use this opcode for now (e.g.
if (oObj <> $) is changed but if (oObj <> $ AND oObj = what) is not.